Fix the default revision for existing/deleted files when parsing DiffX.
An issue was also found where deleted files in a DiffX were getting a
default revision of `PRE_CREATION`, which caused the deleted file to
show up as new. This should have only occurred for newly-created files.

Now, missing revisions show up as `UNKNOWN`, making it more clear that
the revision is missing, and avoiding accidentally triggering this bad
state.

Testing Done:
Unit tests pass on Python 2 and 3.

Tested this in production and verified that it fixed an issue with a
file appearing as new when it should have appeared as deleted.

Reviewed at https://reviews.reviewboard.org/r/11770/