Fix pending/update changeset behavior.
`ReviewRequest.changeset_is_pending` was broken when we did some changes
to the repository/scmtool API, because we neglected to replace the call
to `get_scmtool()`. This wasn't caught by our existing unit test because
the test was broken in a couple ways (it wasn't properly setting
`supports_pending_changesets` on the SCMTool class, and the check for
the new commit ID was overwriting the old value and then testing a
variable against itself).

Testing Done:
Ran the fixed unit test and saw it fail. Applied the fix and saw all
unit tests succeed.

Reviewed at https://reviews.reviewboard.org/r/11795/