Add **kwargs to all unit test model creation helper functions.
Our own `TestCase` class provides a number of helper functions for
easily creating models. Some of these took a `**kwargs` argument, which
was passed in to the model's constructor, allowing arbitrary fields to
be provided, but not all.

This change adds support for `**kwargs` in all the helpers that were
missing it.

Testing Done:
All unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11716/