Switch away from text_type shims.
In the Python 2 to 3 conversion, the name of the binary type was changed
from `str` to `bytes`, and the name of the string type was changed from
`unicode` to `str`. In addition, there were some other changes such as
the removal of the `basestring` type. The `six` library included a bunch
of helpers to ease this transition. Now that we're Python 3+ only on the
master branch, we can get rid of these and just use the builtins.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/11854/