Update the p4python dependency to skip installation on arm64 (Apple M1).
At this time, p4python is not available on Apple M1 systems
(specifically, arm64 systems). There's no way for us to work around
this, and we're hoping that Perforce will address this compatibility
problem in time.

For now, we're skipping the dependency on this system by taking
advantage of the `platform_machine` environment marker.

The comment has also been updated to reflect the curernt version of the
p4python bindings and that version's compatibility.

Testing Done:
Tested installation on a range of Python versions on M1 and on Intel.
The dependency was skipped on M1.

Reviewed at https://reviews.reviewboard.org/r/11873/