Fix name of through table attribute for m2m field.
Django has changed the naming of the related field on `ManyToMany`
fields, renaming it from `related_field` to `rel`. We had one place that
was using this to access the through table, which needed to be updated
for that new name.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12014/