Use display_id when forming URLs in search tests.
The search tests have a bunch of test cases that were querying URLs
based on the `local_id` field. For some of these, the test is checking
that a review request on the global site does not show up when searching
within a local site. In these cases, the `local_id` field can be `None`,
which now causes a reverse lookup error (before it ended up just
sticking "None" in for the ID, which used to give us a URL but didn't
make sense).

I've changed all the test cases to instead use the `display_id` field
when creating API URLs within the tests.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12016/