Update URL patterns to use new path() and re_path().
Django 2+ provides a new method for definings URLs, `path()`. This is
much simpler for basic cases, and makes URL patterns more readable. In
addition, the old `url()` method has been renamed to `re_path()`, and a
warning is raised when using the old name, so switch to that too.

As part of this, I changed the file attachment URLs to actually parse
UUIDs using Django's new converter for that. This required making a
change in the test `create_file_attachment_base` method to use real
UUIDs instead of the string "test-uuid".

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12022/