Call middleware with new style invocation.
The old `process_request` method for middleware was deprecated with the
move to `MIDDLEWARE` from the old `MIDDLEWARE_CLASSES`. Our testcase
`create_http_request` method does a bunch of stuff to build a fake HTTP
request, including running the session and message middleware to
populate attributes on the request. This change moves that invocation
over to the new style (instantiate with a `get_response` method, and
then treat as a callable).

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12027/