Get rid of support for legacy SCMTool method signatures.
SCMTool had a class `__new__` which looked for kwargs support in the
`normalize_path_for_display` method and warned if it wasn't present (and
adapted the method signature with a wrapper). This warned that support
would go away in 5.0, so it's time to remove it.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12032/