Mark Diff ACLs and the DiffX parser as experimental.
Diff ACLs and DiffX support are two major features that are in testing
with customers, and are going to be highlights of the 5.0 release.
Still, we want customers on 4.0 to be able to begin experimenting with
these, and have the freedom to alter the APIs if necessary.

This change marks both as experimental.

For `DiffXParser`, this is just a documentation change, as it's only
used by SCMs that explicitly create it for now (there's no wide-spread
DiffX support across SCMs).

For `FileDiffACLHook` usage, this is walled off by a `diff_acls_feature`
feature check, which is marked as experimental by default. Users will
need to enable this feature if they want to try it for now. We'll mark
it as stable in 5.0.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11986/