Add better version output and remove warnings in rbssh.
This makes a couple of key changes to help diagnose issues in rbssh.

`--version` now shows the Python version and the Review Board package
location, to help ensure that the right version is being run.

Warnings are now disabled at the top of the module, before we import
anything that may warn. This helps avoid some bad output that can occur,
depending on versions of packages and Python, particularly with
Cryptography.

Testing Done:
Verified that this got rid of warnings that messed up test output.

Verified that `--version` showed the correct information.

Reviewed at https://reviews.reviewboard.org/r/11948/