Don't use RequestContext with render_to_string
Django no longer allows passing in a `RequestContext` to the
`render_to_string` method. For the most part we had already migrated to
passing in plain dicts, but we had several template tags that were
trying to use it with the context that they received. In this case, we
can flatten that context to get a plain dict.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12054/