Remove old encoding_list parameter to diffutils methods.
The `encoding_list` parameter was once used to allow us to plumb through
the list of possible encodings for files shown in the diff viewer. This
list is now computed elsewhere, but we kept the old parameters for
compatibility, with a deprecation warning. This change removes those now
that the deprecation cycle is complete.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12122/