Add a missing dependency bump for django_evolution.
Review Board 4.0.5 was meant to require django_evolution>=2.1.3. This
missed the release, so is now scheduled for 4.0.6.

The 4.0.5 release notes have been updated to mention the requirement to
manually upgrade.

Testing Done:
Built the docs, checking the new text and the link to the project on PyPI.

Tested installing with the new dependency bump.

Reviewed at https://reviews.reviewboard.org/r/12049/