Disallow clicking on the "No issues" text in the issue summary table.
The issue summary table had a mouse cursor and click event enabled for
every row, and this included the row with the "There are no issues" text
we show when the active filter turns up 0 results. Clicking this would
take the user to a 404 page with "undefined" at the end of the URL.

This change makes the click handler more specific, only listening to
events on rows with an issue ID set.

The mouse cursor for tabular fields is now the default for text, rather
than the "it's clickable" pointer. This is a more correct default. The
issue summary table specializes that, setting the cursor only for rows
with an issue ID.

Testing Done:
Unit tests pass.

Tested the mouse cursor on issue rows and non-issue rows. Verified that
the cursor only indicated a row was clickable for issue rows.

Tested clicking issue rows and non-issue rows. Verified that the click
was only handled for issue rows.

Reviewed at https://reviews.reviewboard.org/r/12060/