Add a test SCMTool with a DiffX parser.
This adds `TestToolDiffX`, which is a simple Git-based SCMTool that uses
a DiffX parser by default. This allows us to more easily test operations
involving a repository that uses DiffX files.

As of this change, nothing currently uses this tool, but upcoming
changes will begin to make use of it.

Testing Done:
Unit tests pass.

Successfully made use of this with an upcoming change.

Reviewed at https://reviews.reviewboard.org/r/12065/