Fix callback adapter for ready() methods.
The boilerplate that I used for all the callbacks-to-promises
conversions works in almost every case, but the `ready()` method is a
bit special because the callback name is `ready` instead of `success`.
The `promiseToCallbacks` adapter handled this, but the call sites inside
of the `ready` methods weren't testing for that callback name, so it
wouldn't work properly when calling in the old style.

Testing Done:
Used an extension that called `ready()` on a draft review using the
callbacks style instead of promises. Saw that it worked correctly.

Reviewed at https://reviews.reviewboard.org/r/12080/