Replace a use of _super() that wasn't resolving correctly.
Once upon a time, we were trying to make all our backbone "classes" work
more like classes in many other languages, and we added a `_super()`
function which attempts to automatically resolve the extended prototype.
When this fails, it has a habit of infinitely recursing.

I hit this recently attempting to upload a file attachment on
release-5.0.x. I'm not sure if there's a change in the code or a change
with how babel is transpiling something. In these cases, the solution is
to replace the call with an explicit call to the parent prototype
method.

Testing Done:
Uploaded a file attachment and saw the dialog close without any errors
in the JS console.

Reviewed at https://reviews.reviewboard.org/r/12105/