Added styling for logged-out mobile user actions.
The "Log in" and "Register" actions were being displayed as default list
items in the mobile menu.

This adds the `rb-c-page-sidebar__mobile-user-actions` class to the
unordered list for users that are not authenticated so the actions will
be displayed with the proper styling.

Testing Done:
Ran with `./contrib/internal/devserver.py`.

Verified in mobile-view in Chrome that the actions are shown as commands
along the top instead of default list styling as before.

Bugs closed: 4957

Reviewed at https://reviews.reviewboard.org/r/11955/