Add the binary flag to the FileDiff API.
The FileDiff API never exposed the `FileDiff.binary` flag. This
prevented clients from being able to process based on whether this flag
is set.

This change exposes the flag. This will be public as of 4.0.6.

Testing Done:
Unit tests pass.

Verified this flag was shown in the API payloads.

Reviewed at https://reviews.reviewboard.org/r/12094/