Add skipping for two more tests which rely on the P4 module.
There are a couple tests for the repository form which attempt to create
a Perforce-backed repository, but this fails if the P4 module is not
installed. Since there are still a few cases where we're unable to
install the P4 module, these should be marked to skip if it's not
available.

Testing Done:
Ran tests without P4 installed and saw the tests appropriately skipped.

Reviewed at https://reviews.reviewboard.org/r/12116/