Fix up a bad pathological test path.
Inside the API paginator tests, we had a code path to handle
pathological cases where we hit the paginator with an unexpected URL.
This was calling `self.fail()`, but the method this is in isn't a test
case, it's a paginator. This has been changed to just raise an exception
instead.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12115/